Adding more projects
Facebook.Unity.Core will contain things shared by Facebook.Unity that
rely on no internal interfaces.

Facebook.Unity.Internal will contain all internal interfaces.

Reason for doing this is purely because I am not a fan of
InternalsVisibleToAttribute. All this infrastructure that is in place to
use interfaces is ultimately undone by that restriction, such as if
someone wanted to add another platform to support in a separate dll.